fix long

+++ Binomial.java
@@ -18,12 +18,12 @@ public class Binomial {
 	}
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, 
 		//// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
 		//// Replace the following statement with your code.
-		int[][] memo = new int[Math.max(n,k)+1][Math.max(n,k)+1];
+		long[][] memo = new long[Math.max(n,k)+1][Math.max(n,k)+1];
 		for (int i = 0; i < memo.length; i ++) {
 			for (int j = 0; j < memo[0].length; j ++) {
 				memo[i][j] = -1;
@@ -32,7 +32,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
